requirements.txt
setuptools==58.0.0
psutil==5.9.0
torch==1.10.0
torchvision==0.11.1
numpy==1.21.2
matplotlib==3.4.3
requests==2.26.0
pillow==8.4.0
